Update README.md
the 2.0.x.dmg files are using the APFS-type file system and is unmountable by El Capitan, or Sierra. It should say High Sierra or newer because Apple introduced APFS in macOS version 10.13 High Sierra. Running this command on the JabraChromeHost2.0.0.dmg or JabraChromeHost2.0.2.dmg will return that the file system used in the DMG is APFS: hdiutil attach -verbose filepath/to/file.dmg